access denied by user in events controller

+++ src/main/java/com/calendar/services/EventService.java
@@ -1,60 +1,60 @@
 package com.calendar.services;
 
+import com.calendar.exceptions.AccessDeniedForThisUser;
+import com.calendar.exceptions.NotFoundException;
 import com.calendar.models.Event;
-import com.calendar.models.User;
 import com.calendar.repositories.EventRepository;
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
 public class EventService {
     EventRepository repository;
 
-    public List<Event> findAll(){
-        return repository.findAll();
-    }
-
     public List<Event> findByUserId(Long userId){
         return repository.findEventsByUserId(userId);
     }
 
-    public List<Event> findByUserAndStatus(User user, Event.EventStatus status){
-        return repository.findEventsByUserAndStatus(user, status);
-    }
-
     public Event findById(Long id){
-        return repository.findUserById(id);
+        return repository.findEventById(id)
+                .orElseThrow(() -> new NotFoundException("Событие по данному идентификатору не найдено"));
     }
 
-    public List<Event> findByStatus(Event.EventStatus status){
-        return repository.findEventsByStatus(status);
-    }
-
-    public Set<Event.EventStatus> statusList(){
-        return Arrays.stream(Event.EventStatus.values()).collect(Collectors.toSet());
-    }
 
     public Event save(Event event){
         return repository.save(event);
     }
 
-    public Event update(Long id, Event event){
+    public Event update(Long id, Event event, long userId){
+         if(findById(id).getUser().getId() != userId)
+             throw new AccessDeniedForThisUser();
+
          event.setUpdated(LocalDateTime.now());
          return repository.save(event);
     }
 
-    public Event delete(Long id){
+    public Event delete(Long id, long userId){
         Event event = findById(id);
-        if(event != null)
+        if(event != null) {
+            if(event.getUser().getId() != userId) {
+                throw new AccessDeniedForThisUser();
+            }
+
             repository.delete(event);
+        }else
+            throw new NotFoundException("event not found by id");
+
+        return event;
+    }
+
+    public Event findByIdAndUserId(Long id, long userId) {
+        Event event = findById(id);
+        if(event.getUser().getId() != userId)
+            throw new AccessDeniedForThisUser();
 
         return event;
     }
